requirements.txt
django 
pillow 
mysqlclient 
